# openadk/swagger/swagger.yaml
...
swagger: "2.0"
info:
  description: "Yanshee RESTful APIs是一套专门为编程爱好者提供二次开发的接口．它通过http/https的方式向外界提供机器人控\
    制，传感器读取，机器人视觉等功能．用户还可以通过编程的方式训练自己的模型，学习一些机器学习的内容．"
  version: "1.0.0"
  title: "Yanshee RESTful API"
  termsOfService: "https://account.ubtrobot.com/gdpr/ukit/v1.0.3/TermsOfUse_en.html"
  contact:
    email: "swenggroup@ubtrobot.com"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
host: "virtserver.swaggerhub.com"
basePath: "/cygnusnow/OpenADK/1.0.0"
...
schemes:
- "https"
- "http"
paths:
  /subscriptions/motions:
    put:
      tags:
      - "subscriptions"
      summary: "推送运动控制状态"
      description: "可以控制执行指定动作、暂停、继续、停止和复位"
      operationId: "put_motions"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "运动控制的参数"
        required: true
        schema:
          $ref: "#/definitions/MotionsStatusResponse"
      responses:
        200:
          description: "成功"
          schema:
            $ref: "#/definitions/CommonResponse"
      x-swagger-router-controller: "openadk.controllers.subscriptions_controller"
  /subscriptions/sensors/gyro:
    put:
      tags:
      - "subscriptions"
      summary: "推送传感器消息"
      description: ""
      operationId: "put_sensors_subscription"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        required: true
        schema:
          $ref: "#/definitions/SensorsGyroValueResponse"
      responses:
        200:
          description: "发送成功"
          schema:
            $ref: "#/definitions/CommonResponse"
      x-swagger-router-controller: "openadk.controllers.subscriptions_controller"
  /subscriptions/sensors/ultrasonic:
    put:
      tags:
      - "subscriptions"
      summary: "推送传感器消息"
      description: ""
      operationId: "put_sensors_subscription_sensors_ultrasonic"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        required: true
        schema:
          $ref: "#/definitions/SensorsUltrasonicValueResponse"
      responses:
        200:
          description: "发送成功"
          schema:
            $ref: "#/definitions/CommonResponse"
      x-swagger-router-controller: "openadk.controllers.subscriptions_controller"
  /subscriptions/sensors/infrared:
    put:
      tags:
      - "subscriptions"
      summary: "推送传感器消息"
      description: ""
      operationId: "put_sensors_subscription_sensors_infrared"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        required: true
        schema:
          $ref: "#/definitions/SensorsInfraredValueResponse"
      responses:
        200:
          description: "发送成功"
          schema:
            $ref: "#/definitions/CommonResponse"
      x-swagger-router-controller: "openadk.controllers.subscriptions_controller"
  /subscriptions/sensors/environment:
    put:
      tags:
      - "subscriptions"
      summary: "推送传感器消息"
      description: ""
      operationId: "put_sensors_subscription_sensors_environment"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        required: true
        schema:
          $ref: "#/definitions/SensorsEnvironmentValueResponse"
      responses:
        200:
          description: "发送成功"
          schema:
            $ref: "#/definitions/CommonResponse"
      x-swagger-router-controller: "openadk.controllers.subscriptions_controller"
  /subscriptions/sensors/touch:
    put:
      tags:
      - "subscriptions"
      summary: "推送传感器消息"
      description: ""
      operationId: "put_sensors_subscription_sensors_touch"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        required: true
        schema:
          $ref: "#/definitions/SensorsTouchValueResponse"
      responses:
        200:
          description: "发送成功"
          schema:
            $ref: "#/definitions/CommonResponse"
      x-swagger-router-controller: "openadk.controllers.subscriptions_controller"
  /subscriptions/sensors/pressure:
    put:
      tags:
      - "subscriptions"
      summary: "推送传感器消息"
      description: ""
      operationId: "put_sensors_subscription_sensors_pressure"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        required: true
        schema:
          $ref: "#/definitions/SensorsPressureValueResponse"
      responses:
        200:
          description: "发送成功"
          schema:
            $ref: "#/definitions/CommonResponse"
      x-swagger-router-controller: "openadk.controllers.subscriptions_controller"
  /subscriptions/visions:
    put:
      tags:
      - "subscriptions"
      summary: "推送指定视觉任务消息"
      description: "URL example: http://10.10.1.30:80/subscriptions/visions"
      operationId: "put_vision_subscription_visions"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        required: true
        schema:
          $ref: "#/definitions/VisionsResponse"
      responses:
        200:
          description: "发送成功"
          schema:
            $ref: "#/definitions/CommonResponse"
      x-swagger-router-controller: "openadk.controllers.subscriptions_controller"
  /subscriptions/voice/iat:
    put:
      tags:
      - "subscriptions"
      summary: "推送语音识别原始JSON信息"
      description: "URL example: http://10.10.1.30:80/subscriptions/voice/iat"
      operationId: "put_voice_iat_subscription_voice_iat"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        required: true
        schema:
          $ref: "#/definitions/VoiceResponse"
      responses:
        200:
          description: "发送成功"
          schema:
            $ref: "#/definitions/CommonResponse"
      x-swagger-router-controller: "openadk.controllers.subscriptions_controller"
  /subscriptions/voice/asr:
    put:
      tags:
      - "subscriptions"
      summary: "推送语义理解消息"
      description: "URL example: http://10.10.1.30:80/subscriptions/voice/asr"
      operationId: "put_voice_asr_subscriptions_voice_asr"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        required: true
        schema:
          $ref: "#/definitions/VoiceResponse"
      responses:
        200:
          description: "发送成功"
          schema:
            $ref: "#/definitions/CommonResponse"
      x-swagger-router-controller: "openadk.controllers.subscriptions_controller"
  /subscriptions/voice/tts:
    put:
      tags:
      - "subscriptions"
      summary: "推送TTS状态消息"
      description: "URL example: http://10.10.1.30:80/tts"
      operationId: "put_tts_subscriptions_voice_tts"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        required: true
        schema:
          $ref: "#/definitions/CommonResponse"
      responses:
        200:
          description: "发送成功"
          schema:
            $ref: "#/definitions/CommonResponse"
      x-swagger-router-controller: "openadk.controllers.subscriptions_controller"
definitions:
  Name:
    type: "object"
    required:
    - "name"
    properties:
      name:
        type: "string"
    xml:
      name: "Name"
    example:
      name: "name"
...
  CommonResponse:
    type: "object"
    required:
    - "code"
    - "data"
    - "msg"
    properties:
      code:
        type: "integer"
        format: "int32"
        description: "返回码，0表示正常,其它都为错误"
      data:
        type: "object"
        example: {}
        properties: {}
      msg:
        type: "string"
        example: "success"
    xml:
      name: "CommonResponse"
    example:
      msg: "success"
      code: 0
      data: {}
...
  MotionsStatusResponse:
    type: "object"
    required:
    - "status"
    properties:
      status:
        type: "string"
        description: "运行状态"
        enum:
        - "idle"
        - "running"
        - "pause"
      name:
        type: "string"
        description: "动作名"
        default: "none"
      timestamp:
        type: "integer"
        format: "int32"
    xml:
      name: "MotionsStatusResponse"
    example:
      name: "none"
      status: "idle"
      timestamp: 0
...
  SensorsGyroInfo:
    type: "object"
    required:
    - "euler-x"
    - "euler-y"
    - "euler-z"
    - "id"
    properties:
      id:
        type: "integer"
        format: "int32"
        example: 52
        minimum: 1
        maximum: 127
      gyro-x:
        type: "number"
        format: "float"
        example: 0.1021
      gyro-y:
        type: "number"
        format: "float"
        example: 0.602
      gyro-z:
        type: "number"
        format: "float"
        example: 0.0213
      accel-x:
        type: "number"
        format: "float"
        example: 0.147949
      accel-y:
        type: "number"
        format: "float"
        example: 0.116699
      accel-z:
        type: "number"
        example: 0.8134770000000000056417093219351954758167266845703125
      compass-x:
        type: "number"
        format: "float"
        example: 0.1021
      compass-y:
        type: "number"
        format: "float"
        example: 0.602
      compass-z:
        type: "number"
        format: "float"
        example: 0.01232
      euler-x:
        type: "number"
        format: "float"
        example: 8.073578
      euler-y:
        type: "number"
        format: "float"
        example: -10.11795
      euler-z:
        type: "number"
        format: "float"
        example: 0.72052
    xml:
      name: "SensorsGyroInfo"
    example:
      euler-y: -10.11795
      euler-x: 8.073578
      euler-z: 0.72052
      compass-x: 0.1021
      compass-y: 0.602
      compass-z: 0.01232
      gyro-y: 0.602
      accel-y: 0.116699
      id: 52
      gyro-x: 0.1021
      accel-z: 0.8134770000000000056417093219351954758167266845703125
      gyro-z: 0.0213
      accel-x: 0.147949
  SensorsGyroValue:
    type: "object"
    required:
    - "gyro"
    properties:
      gyro:
        type: "array"
        items:
          $ref: "#/definitions/SensorsGyroInfo"
    xml:
      name: "SensorsGyroValue"
    example:
      gyro:
      - euler-y: -10.11795
        euler-x: 8.073578
        euler-z: 0.72052
        compass-x: 0.1021
        compass-y: 0.602
        compass-z: 0.01232
        gyro-y: 0.602
        accel-y: 0.116699
        id: 52
        gyro-x: 0.1021
        accel-z: 0.8134770000000000056417093219351954758167266845703125
        gyro-z: 0.0213
        accel-x: 0.147949
      - euler-y: -10.11795
        euler-x: 8.073578
        euler-z: 0.72052
        compass-x: 0.1021
        compass-y: 0.602
        compass-z: 0.01232
        gyro-y: 0.602
        accel-y: 0.116699
        id: 52
        gyro-x: 0.1021
        accel-z: 0.8134770000000000056417093219351954758167266845703125
        gyro-z: 0.0213
        accel-x: 0.147949
  SensorsGyroValueResponse:
    type: "object"
    required:
    - "code"
    - "data"
    - "msg"
    properties:
      code:
        type: "integer"
        format: "int32"
        description: "返回码，0表示正常，其它值均为错误"
      data:
        $ref: "#/definitions/SensorsGyroValue"
      msg:
        type: "string"
        example: "success"
        description: "返回码的信息"
    xml:
      name: "SensorsGyroValueResponse"
    example:
      msg: "success"
      code: 0
      data:
        gyro:
        - euler-y: -10.11795
          euler-x: 8.073578
          euler-z: 0.72052
          compass-x: 0.1021
          compass-y: 0.602
          compass-z: 0.01232
          gyro-y: 0.602
          accel-y: 0.116699
          id: 52
          gyro-x: 0.1021
          accel-z: 0.8134770000000000056417093219351954758167266845703125
          gyro-z: 0.0213
          accel-x: 0.147949
        - euler-y: -10.11795
          euler-x: 8.073578
          euler-z: 0.72052
          compass-x: 0.1021
          compass-y: 0.602
          compass-z: 0.01232
          gyro-y: 0.602
          accel-y: 0.116699
          id: 52
          gyro-x: 0.1021
          accel-z: 0.8134770000000000056417093219351954758167266845703125
          gyro-z: 0.0213
          accel-x: 0.147949
  SensorsEnvironmentInfo:
    type: "object"
    required:
    - "humidity"
    - "id"
    - "temperature"
    properties:
      id:
        type: "integer"
        format: "int32"
        example: 54
        minimum: 1
        maximum: 127
      temperature:
        type: "integer"
        format: "int32"
        example: 25
      humidity:
        type: "integer"
        format: "int32"
        example: 75
      pressure:
        type: "integer"
        format: "int32"
        example: 1003
    xml:
      name: "SensorsEnvironmentInfo"
    example:
      temperature: 25
      humidity: 75
      id: 54
      pressure: 1003
  SensorsEnvironmentValue:
    type: "object"
    required:
    - "environment"
    properties:
      environment:
        type: "array"
        items:
          $ref: "#/definitions/SensorsEnvironmentInfo"
    xml:
      name: "SensorsEnvironmentValue"
    example:
      environment:
      - temperature: 25
        humidity: 75
        id: 54
        pressure: 1003
      - temperature: 25
        humidity: 75
        id: 54
        pressure: 1003
  SensorsEnvironmentValueResponse:
    type: "object"
    required:
    - "code"
    - "data"
    - "msg"
    properties:
      code:
        type: "integer"
        format: "int32"
        description: "返回码，0表示正常，其它值均为错误"
      data:
        $ref: "#/definitions/SensorsEnvironmentValue"
      msg:
        type: "string"
        example: "success"
        description: "返回码的信息"
    xml:
      name: "SensorsEnvironmentValueResponse"
    example:
      msg: "success"
      code: 0
      data:
        environment:
        - temperature: 25
          humidity: 75
          id: 54
          pressure: 1003
        - temperature: 25
          humidity: 75
          id: 54
          pressure: 1003
  SensorsCommonInfo:
    type: "object"
    required:
    - "id"
    - "value"
    properties:
      id:
        type: "integer"
        format: "int32"
        example: 16
        minimum: 1
        maximum: 127
      value:
        type: "integer"
        format: "int32"
        example: 25
    xml:
      name: "SensorsCommonInfo"
    example:
      id: 16
      value: 25
  SensorsUltrasonicValue:
    type: "object"
    required:
    - "ultrasonic"
    properties:
      ultrasonic:
        type: "array"
        items:
          $ref: "#/definitions/SensorsCommonInfo"
    xml:
      name: "SensorsUltrasonicValue"
    example:
      ultrasonic:
      - id: 16
        value: 25
      - id: 16
        value: 25
  SensorsUltrasonicValueResponse:
    type: "object"
    required:
    - "code"
    - "data"
    - "msg"
    properties:
      code:
        type: "integer"
        format: "int32"
        description: "返回码，0表示正常，其它值均为错误"
      data:
        $ref: "#/definitions/SensorsUltrasonicValue"
      msg:
        type: "string"
        example: "success"
        description: "返回码的信息"
    xml:
      name: "SensorsUltrasonicValueResponse"
    example:
      msg: "success"
      code: 0
      data:
        ultrasonic:
        - id: 16
          value: 25
        - id: 16
          value: 25
  SensorsInfraredValue:
    type: "object"
    required:
    - "infrared"
    properties:
      infrared:
        type: "array"
        items:
          $ref: "#/definitions/SensorsCommonInfo"
    xml:
      name: "SensorsInfraredValue"
    example:
      infrared:
      - id: 16
        value: 25
      - id: 16
        value: 25
  SensorsInfraredValueResponse:
    type: "object"
    required:
    - "code"
    - "data"
    - "msg"
    properties:
      code:
        type: "integer"
        format: "int32"
        description: "返回码，0表示正常，其它值均为错误"
      data:
        $ref: "#/definitions/SensorsInfraredValue"
      msg:
        type: "string"
        example: "success"
        description: "返回码的信息"
    xml:
      name: "SensorsInfraredValueResponse"
    example:
      msg: "success"
      code: 0
      data:
        infrared:
        - id: 16
          value: 25
        - id: 16
          value: 25
  SensorsTouchValue:
    type: "object"
    required:
    - "touch"
    properties:
      touch:
        type: "array"
        items:
          $ref: "#/definitions/SensorsCommonInfo"
    xml:
      name: "SensorsTouchValue"
    example:
      touch:
      - id: 16
        value: 25
      - id: 16
        value: 25
  SensorsTouchValueResponse:
    type: "object"
    required:
    - "code"
    - "data"
    - "msg"
    properties:
      code:
        type: "integer"
        format: "int32"
        description: "返回码，0表示正常，其它值均为错误"
      data:
        $ref: "#/definitions/SensorsTouchValue"
      msg:
        type: "string"
        example: "success"
        description: "返回码的信息"
    xml:
      name: "SensorsTouchValueResponse"
    example:
      msg: "success"
      code: 0
      data:
        touch:
        - id: 16
          value: 25
        - id: 16
          value: 25
  SensorsPressureValue:
    type: "object"
    required:
    - "pressure"
    properties:
      pressure:
        type: "array"
        items:
          $ref: "#/definitions/SensorsCommonInfo"
    xml:
      name: "SensorsUltrasonicValue"
    example:
      pressure:
      - id: 16
        value: 25
      - id: 16
        value: 25
  SensorsPressureValueResponse:
    type: "object"
    required:
    - "code"
    - "data"
    - "msg"
    properties:
      code:
        type: "integer"
        format: "int32"
        description: "返回码，0表示正常，其它值均为错误"
      data:
        $ref: "#/definitions/SensorsPressureValue"
      msg:
        type: "string"
        example: "success"
        description: "返回码的信息"
    xml:
      name: "SensorsPressureValueResponse"
    example:
      msg: "success"
      code: 0
      data:
        pressure:
        - id: 16
          value: 25
        - id: 16
          value: 25
...
  VoiceResponse:
    type: "object"
    required:
    - "code"
    - "msg"
    - "type"
    properties:
      code:
        type: "integer"
        example: 0
        description: "错误码"
      type:
        type: "string"
        example: "asr"
        description: "消息类型。此处的参数有 'asr', 'iat' "
        enum:
        - "asr"
        - "iat"
      data:
        type: "string"
        description: "数据"
      msg:
        type: "string"
        description: "错误码消息"
    xml:
      name: "VoiceResponse"
    description: "asr与iat推送JSON"
    example:
      msg: "msg"
      code: 0
      data: "data"
      type: "asr"
  VisionsQuantity:
    type: "object"
    required:
    - "quantity"
    properties:
      quantity:
        type: "integer"
        example: 1
        description: "数量, 0 ~ ??? (整数)"
        minimum: 0
    xml:
      name: "VisionsQuantity"
    example:
      quantity: 1
  VisionsGender:
    type: "object"
    required:
    - "gender"
    properties:
      gender:
        type: "string"
        description: "性别。此处的值只有 'male', 'female'"
        enum:
        - "male"
        - "female"
    xml:
      name: "VisionsGender"
    example:
      gender: "male"
  VisionsAge:
    type: "object"
    required:
    - "age"
    properties:
      age:
        type: "integer"
        example: 5
        description: "年龄"
        minimum: 1
    xml:
      name: "VisionsAge"
    example:
      age: 5
  VisionsExpression:
    type: "object"
    required:
    - "anger"
    - "fear"
    - "happiness"
    - "neutral"
    - "sadness"
    - "surprise"
    properties:
      happiness:
        type: "number"
        format: "float"
        example: 86.5
        description: "快乐值, 0 ~ 100 (浮点数)"
        minimum: 0
        maximum: 100
      surprise:
        type: "number"
        format: "float"
        example: 40.7
        description: "吃惊值, 0 ~ 100 (浮点数)"
        minimum: 0
        maximum: 100
      anger:
        type: "number"
        format: "float"
        example: 60.7
        description: "生气值, 0 ~ 100 (浮点数)"
        minimum: 0
        maximum: 100
      sadness:
        type: "number"
        format: "float"
        example: 54.2
        description: "伤心值, 0 ~ 100 (浮点数)"
        minimum: 0
        maximum: 100
      neutral:
        type: "number"
        format: "float"
        example: 53.6
        description: "平静值, 0 ~ 100 (浮点数)"
        minimum: 0
        maximum: 100
      fear:
        type: "number"
        format: "float"
        example: 42.8
        description: "害怕值, 0 ~ 100 (浮点数)"
        minimum: 0
        maximum: 100
    xml:
      name: "VisionsExpression"
    example:
      surprise: 40.7
      happiness: 86.5
      neutral: 53.6
      sadness: 54.2
      anger: 60.7
      fear: 42.8
  VisionsAnalysis:
    type: "object"
    properties:
      age:
        $ref: "#/definitions/VisionsAge"
      gender:
        $ref: "#/definitions/VisionsGender"
      expression:
        $ref: "#/definitions/VisionsExpression"
    xml:
      name: "VisionsAnalysis"
    example:
      expression:
        surprise: 40.7
        happiness: 86.5
        neutral: 53.6
        sadness: 54.2
        anger: 60.7
        fear: 42.8
      gender:
        gender: "male"
      age:
        age: 5
  VisionsResponseResponse:
    type: "object"
    properties:
      analysis:
        $ref: "#/definitions/VisionsAnalysis"
      recognition:
        $ref: "#/definitions/Name"
      quantity:
        $ref: "#/definitions/VisionsQuantity"
    xml:
      name: "VisionsResponseResponse"
    example:
      quantity:
        quantity: 1
      recognition:
        name: "name"
      analysis:
        expression:
          surprise: 40.7
          happiness: 86.5
          neutral: 53.6
          sadness: 54.2
          anger: 60.7
          fear: 42.8
        gender:
          gender: "male"
        age:
          age: 5
  VisionsResponse:
    type: "object"
    required:
    - "code"
    - "msg"
    - "type"
    properties:
      code:
        type: "integer"
        example: 0
        description: "错误码"
      type:
        type: "string"
        description: "消息类型。 一次只返回一种类型的数据。 type 允许的值为: 'idle, recognition', 'tracking',\
          \ 'analysis', 'quantity'"
        enum:
        - "recognition"
        - "tracking"
        - "age_analysis"
        - "gender_analysis"
        - "expression_analysis"
        - "quantity"
      data:
        $ref: "#/definitions/VisionsResponseResponse"
      msg:
        type: "string"
        example: "success"
        description: "错误码消息"
    xml:
      name: "VisionsResponse"
    example:
      msg: "success"
      code: 0
      data:
        quantity:
          quantity: 1
        recognition:
          name: "name"
        analysis:
          expression:
            surprise: 40.7
            happiness: 86.5
            neutral: 53.6
            sadness: 54.2
            anger: 60.7
            fear: 42.8
          gender:
            gender: "male"
          age:
            age: 5
      type: "recognition"
...